# antibody-api-spec.yaml
openapi: 3.0.0
info:
  description: 'The HuBMAP Antibody API is a standard RESTful web service with create operations.<br><br>'
  version: 1.0.0
  title: HuBMAP Antibody API
  contact:
    name: HuBMAP Help Desk
    email: api-developers@hubmapconsortium.org
  license:
    name: MIT License
    url: 'https://github.com/hubmapconsortium/antibody-api/blob/master/license.txt'
servers:
  - url: 'https://avr.hubmapconsortium.org/'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Antibody:
      type: object
      properties:
        protocol_doi:
          type: string
          description: "The protocols.io digital object identifier for the protocol under which the sample was obtained and/or prepared."
        uniprot_accession_number:
          type: string
          description: "https://www.uniprot.org/help/accession_numbers"
        target_symbol:
          type: string
          description: "Either a HGNC Approved Symbol, Previous Symbol or Alias Symbol which is resolved into an Approved Symbol."
        rrid:
          type: string
          description: "The associated Research Resource IDentifier (https://scicrunch.org/resources)."
        host_organism:
          type: string
          description: "The antibody's host organism."
        clonality:
          type: string
          enum:
            - monoclonal
            - polyclonal
            - oligoclonal
          description: "One of the values: monoclonal (one), oligoclonal (few), and polyclonal (many)."
        vendor:
          type: string
          description: "The antibody's vendor."
        catalog_number:
          type: string
        lot_number:
          type: string
        recombinant:
          type: boolean
          description: "True if the antibody is recombinant."
        organ_or_tissue:
          type: string
          description: "Organ or tissue affected."
        hubmap_platform:
          type: string
          description: "The assay employed for the antibody."
        submitter_orcid:
          type: string
          description: "The ORCID iD of the submitter (https://orcid.org/)"
    AntibodyListed:
      allOf:
        - $ref: '#/components/schemas/Antibody'
        - type: object
          properties:
            antibody_uuid:
              type: string
              description: "The unique identifier for the antibody."
            created_by_user_displayname:
              type: string
              description: "The name of the person or process authenticated when creating the object."
            created_by_user_email:
              type: string
              description: "The email address of the person or process authenticated when creating the object."
            created_by_user_sub:
              type: string
              description: "The subject id as provided by the authorization mechanism for the person or process authenticated when creating the object."
            group_uuid:
              type: string
              description: "The unique identifier for the group for the person or process authenticated when creating the object."
security:
  - bearerAuth: []
paths:
  '/antibodies':
    get:
      summary: List antibodies
      responses:
        '200':
          description: List the existing antibodies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AntibodyListed'
        '500':
          description: Internal error
    post:
      summary: Save antibody
      parameters:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                antibody:
                  $ref: '#/components/schemas/Antibody'
                  description: "The Antibody to be saved."
      responses:
        '200':
          description: The antibody was successfully created and the new ID is returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: "The ID of the Antibody just created."
        '400':
          description: The antibody data submitted did not pass validation
        '401':
          description: The user's token has expired or the user did not supply a valid token
        '403':
          description: The user is not authorized to create the entity
        '406':
          description: No JSON body supplied
        '500':
          description: Internal error
  '/restore_elasticsearch':
    put:
      summary: Restore Elastic Search from Database
      responses:
        '200':
          description: Success
        '500':
          description: Internal error